+++ bin/network-monitor/src/config.rs
@@ -13,8 +13,6 @@ use url::Url;
 // ================================================================================================
 
 const DEFAULT_RPC_URL: &str = "http://0.0.0.0:57291";
-const DEFAULT_REMOTE_PROVER_URLS: &str = "http://localhost:50052";
-const DEFAULT_FAUCET_URL: &str = "http://localhost:8080";
 const DEFAULT_PORT: u16 = 3000;
 
 /// Configuration for the monitor.
@@ -35,7 +33,6 @@ pub struct MonitorConfig {
     #[arg(
         long = "remote-prover-urls",
         env = "MIDEN_MONITOR_REMOTE_PROVER_URLS",
-        default_value = DEFAULT_REMOTE_PROVER_URLS,
         value_delimiter = ',',
         help = "The URLs of the remote provers for status checking (comma-separated)"
     )]
@@ -45,7 +42,6 @@ pub struct MonitorConfig {
     #[arg(
         long = "faucet-url",
         env = "MIDEN_MONITOR_FAUCET_URL",
-        default_value = DEFAULT_FAUCET_URL,
         help = "The URL of the faucet service for testing (optional)"
     )]
     pub faucet_url: Option<Url>,
@@ -100,6 +96,29 @@ pub struct MonitorConfig {
     )]
     pub enable_otel: bool,
 
+    /// Whether to disable the network transaction service checks (enabled by default). The network
+    /// transaction service is a network account with a counter deployed at startup and incremented
+    /// by sending a transaction to it.
+    #[arg(
+        long = "disable-ntx-service",
+        env = "MIDEN_MONITOR_DISABLE_NTX_SERVICE",
+        action = clap::ArgAction::SetTrue,
+        default_value_t = false,
+        help = "Whether to disable the network transaction service checks (enabled by default). The
+        network transaction service is a network account with a counter deployed at startup and
+        incremented by sending a transaction to it."
+    )]
+    pub disable_ntx_service: bool,
+
+    /// Path for the counter program network account file.
+    #[arg(
+        long = "counter-filepath",
+        env = "MIDEN_MONITOR_COUNTER_FILEPATH",
+        default_value = "counter_program.mac",
+        help = "Path where the counter account is located"
+    )]
+    pub counter_filepath: PathBuf,
+
     /// Path for the wallet account file.
     #[arg(
         long = "wallet-filepath",
@@ -109,12 +128,13 @@ pub struct MonitorConfig {
     )]
     pub wallet_filepath: PathBuf,
 
-    /// Path for the counter program account file.
+    /// The interval at which to send the increment counter transaction.
     #[arg(
-        long = "counter-filepath",
-        env = "MIDEN_MONITOR_COUNTER_FILEPATH",
-        default_value = "counter_program.mac",
-        help = "Path where the counter account is located"
+        long = "counter-increment-interval",
+        env = "MIDEN_MONITOR_COUNTER_INCREMENT_INTERVAL",
+        default_value = "30s",
+        value_parser = humantime::parse_duration,
+        help = "The interval at which to send the increment counter transaction"
     )]
-    pub counter_filepath: PathBuf,
+    pub counter_increment_interval: Duration,
 }